Rewrite in as Bitwig Java Extension
This commit replaces the old Bitwig API v1 controller script with a
Bitwig Extension that uses the v2 API and is written in Java.
It also sets up a CI job that automatically builds and uploads releases.

Things that are currently not mapped or unfunctional are the 9th knob and the drum pads.

A new feature is that presets can be switched with the buttons below the
crossfader where the left button selects the previous preset, the right
button the next one and the middle button commits the preset change.